Move Imgui parts to new debug_ui module
