+++ src/debug_ui.rs
@@ -1,0 +1,89 @@
+use gfx_core::memory::Typed;
+use ggez::*;
+use ggez::event::winit_event::Event;
+use imgui::{Context as ImguiContext, Io, Ui};
+use imgui_gfx_renderer::{Renderer as ImguiRenderer, Shaders};
+use imgui_winit_support::{HiDpiMode, WinitPlatform};
+
+pub struct DebugUi {
+    imgui: ImguiContext,
+    platform: WinitPlatform,
+    renderer: ImguiRenderer<gfx::format::Rgba8, gfx_device_gl::Resources>,
+}
+
+impl DebugUi {
+    pub fn new(ctx: &mut Context) -> Self {
+        // Initialize Dear Imgui and its GFX renderer.
+        let mut imgui = ImguiContext::create();
+        // TODO: Set imgui.ini file path.
+        let mut platform = WinitPlatform::init(&mut imgui);
+        platform.attach_window(imgui.io_mut(), graphics::window(ctx), HiDpiMode::Rounded);
+        let renderer = {
+            let (factory, device, _encoder, _depth_stencil_view, _render_target_view) =
+                graphics::gfx_objects(ctx);
+            let version = device.get_info().shading_language;
+            let shaders = if version.is_embedded {
+                if version.major >= 3 {
+                    Shaders::GlSlEs300
+                } else {
+                    Shaders::GlSlEs100
+                }
+            } else if version.major >= 4 {
+                Shaders::GlSl400
+            } else if version.major >= 3 {
+                if version.minor >= 2 {
+                    Shaders::GlSl150
+                } else {
+                    Shaders::GlSl130
+                }
+            } else {
+                Shaders::GlSl110
+            };
+            ImguiRenderer::init(&mut imgui, factory, shaders)
+                .expect("Could not initialize imgui_gfx_renderer::Renderer")
+        };
+
+        Self {
+            imgui,
+            platform,
+            renderer,
+        }
+    }
+
+    pub fn handle_event(&mut self, ctx: &Context, event: &Event) {
+        self.platform.handle_event(self.imgui.io_mut(), graphics::window(ctx), event);
+    }
+
+    pub fn prepare_frame(&mut self, ctx: &Context) -> GameResult {
+        self.platform.prepare_frame(self.imgui.io_mut(), graphics::window(ctx))
+            .map_err(|err| GameError::WindowError(err))?;
+        // NOTE: Setting delta time directly instead of calling io_mut().update_delta_time().
+        self.imgui.io_mut().delta_time = timer::duration_to_f64(timer::delta(ctx)) as f32;
+
+        Ok(())
+    }
+
+    pub fn draw_ui<F>(&mut self, ctx: &mut Context, func: F)
+        where F: FnOnce(&Ui)
+    {
+        // Build our Dear Imgui UI elements.
+        let ui = self.imgui.frame();
+
+        func(&ui);
+
+        // Render our Dear Imgui UI elements.
+        self.platform.prepare_render(&ui, graphics::window(ctx));
+        let ui_draw_data = ui.render();
+        {
+            let (factory, _device, encoder, _depth_stencil_view, render_target_view) =
+                graphics::gfx_objects(ctx);
+            let mut target = gfx_core::handle::RenderTargetView::new(render_target_view);
+            self.renderer.render(factory, encoder, &mut target, ui_draw_data)
+                .expect("Could not render Dear Imgui UI");
+        }
+    }
+
+    pub fn io(&self) -> &Io {
+        self.imgui.io()
+    }
+}
